+++ core/src/com/unciv/logic/map/mapunit/MapUnitCache.kt
@@ -70,6 +70,9 @@ class MapUnitCache(private val mapUnit: MapUnit) {
     var hasStrengthBonusInRadiusUnique = false
 
     var hasCitadelPlacementUnique = false
+    
+    var distanceToClosestEnemyUnit: Int? = null
+    var distanceToClosestEnemyUnitSearched: Int? = null
 
     fun updateUniques(){