+++ core/src/com/unciv/ui/screens/worldscreen/mainmenu/WorldScreenMenuPopup.kt
@@ -5,7 +5,6 @@ import com.unciv.ui.components.input.KeyboardBinding
 import com.unciv.ui.popups.Popup
 import com.unciv.ui.screens.civilopediascreen.CivilopediaScreen
 import com.unciv.ui.screens.savescreens.LoadGameScreen
-import com.unciv.ui.screens.savescreens.SaveGameScreen
 import com.unciv.ui.screens.victoryscreen.VictoryScreen
 import com.unciv.ui.screens.worldscreen.WorldScreen
 
@@ -21,10 +20,11 @@ class WorldScreenMenuPopup(val worldScreen: WorldScreen) : Popup(worldScreen, sc
             close()
             worldScreen.game.pushScreen(CivilopediaScreen(worldScreen.gameInfo.ruleset))
         }.row()
-        addButton("Save game", KeyboardBinding.SaveGame) {
-            close()
-            worldScreen.game.pushScreen(SaveGameScreen(worldScreen.gameInfo))
-        }.row()
+        if (!worldScreen.gameInfo.gameParameters.isOnlineMultiplayer)
+            addButton("Save game", KeyboardBinding.SaveGame) {
+                close()
+                worldScreen.openSaveGameScreen()
+            }.row()
         addButton("Load game", KeyboardBinding.LoadGame) {
             close()
             worldScreen.game.pushScreen(LoadGameScreen())